requirements.txt
pillow
mutagen
pygame
lyricsgenius
arabic_reshaper
python-bidi